requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
supabase==2.0.3
python-dotenv==1.0.0
anthropic==0.8.1
deep-translator==1.11.4
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
httpx==0.25.2
pydantic==2.5.0
resend==0.5.1
python-multipart==0.0.6
starlette==0.27.0
slowapi==0.1.9
limits==3.6.0
email-validator==2.1.0
python-dateutil==2.8.2
numpy==1.26.2
scikit-learn==1.3.2
nltk==3.8.1
textblob==0.17.1
langdetect==1.0.9
schedule==1.2.1
aiofiles==23.2.1
orjson==3.9.10
apscheduler==3.10.4
python-slugify==8.0.1
